Revised ConvertSegment command and Schema converter
Druid ingestion spec does not have enough info to generate a Pinot schema,
so the schema is now generated directly from the segment by creating
an adapter. Also combined the Schema converter with the ConvertSegment
command instead of having it as a separate command.
