+++ pinot-tools/pom.xml
@@ -123,10 +123,6 @@
         </exclusion>
       </exclusions>
     </dependency>
-    <dependency>
-      <groupId>org.apache.pinot</groupId>
-      <artifactId>pinot-orc</artifactId>
-    </dependency>
   </dependencies>
   <build>
     <plugins>
