+++ pom.xml
@@ -330,7 +330,7 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
           </exclusion>
-      </exclusions>
+        </exclusions>
       </dependency>
       <dependency>
         <groupId>com.google.guava</groupId>
